Change all defaults to float64 for calculations

+++ post_data.go
@@ -3,29 +3,29 @@ package fbintegration
 type (
 	// PostData comment pending
 	PostData struct {
-		Impressions                   int              `json:"impressions"`
-		PaidImpressions               int              `json:"paid_impressions"`
-		OrganicImpressions            int              `json:"organic_impressions"`
-		Reach                         int              `json:"reach"`
-		PaidReach                     int              `json:"paid_reach"`
-		OrganicReach                  int              `json:"organic_reach"`
-		VideoViews                    int              `json:"video_views"`
-		PaidVideoViews                int              `json:"paid_video_views"`
-		OrganicVideoViews             int              `json:"organic_video_views"`
-		UniqueVideoViews              int              `json:"unique_video_views"`
-		MinutesViewed                 int              `json:"minutes_viewed"`
-		AverageDuration               int              `json:"average_duration"`
-		ReactionsTotal                int              `json:"reactions_total"`
-		Reactions                     map[string]int   `json:"reactions"`
-		Comments                      int              `json:"comments"`
-		Shares                        int              `json:"shares"`
-		Clicks                        int64            `json:"clicks"`
-		UniquePeopleEngaged           int64            `json:"unique_people_engaged"`
-		Spend                         float64          `json:"spend"`
-		VideoViewCost                 float64          `json:"video_view_cost"`
-		Actions                       int64            `json:"actions"`
-		EngagementPercentPeopleViewed float64          `json:"engagement_percent_people_viewed"`
-		ViewRate                      float64          `json:"view_rate"`
-		AudienceSplit                 []*AudienceSplit `json:"audience_split"`
+		Impressions                   float64            `json:"impressions"`
+		PaidImpressions               float64            `json:"paid_impressions"`
+		OrganicImpressions            float64            `json:"organic_impressions"`
+		Reach                         float64            `json:"reach"`
+		PaidReach                     float64            `json:"paid_reach"`
+		OrganicReach                  float64            `json:"organic_reach"`
+		VideoViews                    float64            `json:"video_views"`
+		PaidVideoViews                float64            `json:"paid_video_views"`
+		OrganicVideoViews             float64            `json:"organic_video_views"`
+		UniqueVideoViews              float64            `json:"unique_video_views"`
+		MinutesViewed                 float64            `json:"minutes_viewed"`
+		AverageDuration               float64            `json:"average_duration"`
+		ReactionsTotal                float64            `json:"reactions_total"`
+		Reactions                     map[string]float64 `json:"reactions"`
+		Comments                      float64            `json:"comments"`
+		Shares                        float64            `json:"shares"`
+		Clicks                        float64            `json:"clicks"`
+		UniquePeopleEngaged           float64            `json:"unique_people_engaged"`
+		Spend                         float64            `json:"spend"`
+		VideoViewCost                 float64            `json:"video_view_cost"`
+		Actions                       float64            `json:"actions"`
+		EngagementPercentPeopleViewed float64            `json:"engagement_percent_people_viewed"`
+		ViewRate                      float64            `json:"view_rate"`
+		AudienceSplit                 []*AudienceSplit   `json:"audience_split"`
 	}
 )
